+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Karafka framework changelog
 
+## 2.1.13 (Unreleased)
+- **[Feature]** Introduce Cleaning API for much better memory management for iterative data processing [Pro].
+- [Enhancement] Automatically free message resources after processed for ActiveJob jobs [Pro]
+- [Enhancement] Free memory used by the raw payload as fast as possible after obtaining it from `karafka-rdkafka`.
+
 ## 2.1.12 (2023-08-25)
 - [Fix] Fix a case where DLQ + VP without intermediate marking would mark earlier message then the last one.